refactor: Rename `lib/plugins/aws/utils/arnRegularExpressions.js`
