+++ lib/plugins/aws/package/compile/events/apiGateway/lib/method/authorization.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const _ = require('lodash');
-const awsArnRegExs = require('../../../../../../utils/arnRegularExpressions');
+const awsArnRegExs = require('../../../../../../utils/arn-regular-expressions');
 
 module.exports = {
   getMethodAuthorization(http) {
