+++ lib/plugins/aws/package/compile/events/apiGateway/lib/permissions.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const awsArnRegExs = require('../../../../../utils/arnRegularExpressions');
+const awsArnRegExs = require('../../../../../utils/arn-regular-expressions');
 
 module.exports = {
   compilePermissions() {
