+++ lib/plugins/aws/package/compile/events/apiGateway/lib/validate.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const _ = require('lodash');
-const awsArnRegExs = require('../../../../../utils/arnRegularExpressions');
+const awsArnRegExs = require('../../../../../utils/arn-regular-expressions');
 const resolveLambdaTarget = require('../../../../../utils/resolveLambdaTarget');
 const ServerlessError = require('../../../../../../../serverless-error');
 const { log } = require('@serverless/utils/log');
