+++ lib/plugins/aws/utils/arn-regular-expressions.js
